+++ common/src/commonMain/kotlin/pl/karol202/sciorder/common/model/Order.kt
@@ -3,8 +3,8 @@ package pl.karol202.sciorder.common.model
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class Order(override val id: Int,
-                 val storeId: Int,
+data class Order(override val id: Long,
+                 val storeId: Long,
                  val entries: List<Entry>,
                  val details: Details,
                  val status: Status) : JvmSerializable, IdProvider
@@ -12,9 +12,9 @@ data class Order(override val id: Int,
 	companion object;
 
 	@Serializable
-	data class Entry(val productId: Int,
+	data class Entry(val productId: Long,
 	                 val quantity: Int,
-	                 val parameters: Map<Int, String>) : JvmSerializable
+	                 val parameters: Map<Long, String>) : JvmSerializable
 
 	@Serializable
 	data class Details(val location: String,
