+++ common/src/commonMain/kotlin/pl/karol202/sciorder/common/model/Store.kt
@@ -3,5 +3,5 @@ package pl.karol202.sciorder.common.model
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class Store(override val id: Int,
+data class Store(override val id: Long,
                  val name: String) : JvmSerializable, IdProvider
